Completed settings screen. Created SettingsManager. Create optional title screen
Completed settings activity to allow all activities that have a settings
button to access it. Made the settings activity modular so over time
more settings options fragments can be added to the ScrollView. Created
a settings manager to be able to deal with any changes in the settings
and allow the program to easily get settings states instead of directly
from database. Created an optional title screen which displays basic
user information (defaults to don't show title screen). Ready for
release